+++ src/ts/VAST/AdUnits/VastAdUnitFactory.ts
@@ -24,7 +24,7 @@ export class VastAdUnitFactory extends AbstractAdUnitFactory<VastCampaign, IVast
         const vastOverlayHandler = new VastOverlayEventHandler(vastAdUnit, parameters);
         parameters.overlay.addEventHandler(vastOverlayHandler);
 
-        if(parameters.campaign.hasEndscreen() && parameters.endScreen) {
+        if(parameters.campaign.hasStaticEndscreen() && parameters.endScreen) {
             const vastEndScreenHandler = new VastEndScreenEventHandler(vastAdUnit, parameters);
             parameters.endScreen.addEventHandler(vastEndScreenHandler);